Add option to read message from file
This patch adds a option to read the displayed message from a file.
Additionally the maximum length for the read string can be defined.
The string will then be cutted after the given number of chars.
It is also possible to define a prefix which is displayed in front
of the read file content.

If no file is defined a hardcoded message (using MSG like before)
can be displayed.

For these changes the following defines were introduced:
MSG_FILE_PATH ...... path to the file
MSG_FILE_MAX_LEN ... number of chars read from file
MSG_FILE_PREFIX .... text displayed in front of the file content

Signed-off-by: Richard Leitner <richard.leitner@skidata.com>