add recovered bar.png
While the source files for the main splash image is present in
the source folder base-image, the progress bar isn't. This
patch adds the bar.png recovered from the RLE data in the bar
header file. The tool make-image-header.sh allows to translate
this png back the the header exactly.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>